Move redux entities helpers to store/entities/utils
The moved modules contain helper functions/classes for the to be created
entities actions and selectors. Therefore use a more obvious directory
scheme.
