+++ gsa/CMakeLists.txt
@@ -763,12 +763,12 @@ set (GSA_JS_SRC_FILES
      ${GSA_SRC_DIR}/src/web/store/dashboard/settings/selectors.js
      ${GSA_SRC_DIR}/src/web/store/index.js
      ${GSA_SRC_DIR}/src/web/store/reducers.js
-     ${GSA_SRC_DIR}/src/web/store/entities/actions.js
-     ${GSA_SRC_DIR}/src/web/store/entities/selectors.js
      ${GSA_SRC_DIR}/src/web/store/entities/filters/actions.js
      ${GSA_SRC_DIR}/src/web/store/entities/filters/reducers.js
      ${GSA_SRC_DIR}/src/web/store/entities/filters/selectors.js
      ${GSA_SRC_DIR}/src/web/store/entities/reducers.js
+     ${GSA_SRC_DIR}/src/web/store/entities/utils/actions.js
+     ${GSA_SRC_DIR}/src/web/store/entities/utils/selectors.js
      ${GSA_SRC_DIR}/src/web/utils/cert.js
      ${GSA_SRC_DIR}/src/web/utils/compose.js
      ${GSA_SRC_DIR}/src/web/utils/cpe.js
