+++ gsa/src/web/store/entities/filters/selectors.js
@@ -20,7 +20,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-import EntitiesSelector from '../selectors';
+import EntitiesSelector from '../utils/selectors';
 
 const getFilters = rootState => {
   return new EntitiesSelector(rootState.entities.filters);
