+++ gsa/src/web/store/entities/utils/__tests__/actions.js
